Removing the main legacy checkout backend class
