+++ changelog.txt
@@ -1,6 +1,7 @@
 *** Changelog ***
 
 = 10.3.0 - xxxx-xx-xx =
+* Update - Removes the legacy checkout experience gateway class
 * Update - Remove legacy checkout payment method classes
 * Dev - Renames the express checkout frontend main setting key
 * Update - Deprecates and replaces Payment Request Button classes with Express Checkout equivalents
