+++ includes/payment-methods/class-wc-stripe-upe-payment-method-amazon-pay.php
@@ -111,11 +111,11 @@ class WC_Stripe_UPE_Payment_Method_Amazon_Pay extends WC_Stripe_UPE_Payment_Meth
 	/**
 	 * Return if Amazon Pay is enabled.
 	 *
-	 * @param WC_Gateway_Stripe $gateway The gateway instance.
+	 * @param WC_Stripe_Payment_Gateway $gateway The gateway instance.
 	 *
 	 * @return bool
 	 */
-	public static function is_amazon_pay_enabled( WC_Gateway_Stripe $gateway ) {
+	public static function is_amazon_pay_enabled( WC_Stripe_Payment_Gateway $gateway ) {
 		// Amazon Pay is disabled if feature flag is disabled.
 		if ( ! WC_Stripe_Feature_Flags::is_amazon_pay_available() ) {
 			return false;
