+++ readme.txt
@@ -140,6 +140,7 @@ If you get stuck, you can ask for help in the [Plugin Forum](https://wordpress.o
 == Changelog ==
 
 = 10.3.0 - xxxx-xx-xx =
+* Update - Removes the legacy checkout experience gateway class
 * Update - Remove legacy checkout payment method classes
 * Dev - Renames the express checkout frontend main setting key
 * Update - Deprecates and replaces Payment Request Button classes with Express Checkout equivalents
